% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{Safely convert to numeric}
\usage{
as_numeric(.x)
}
\arguments{
\item{.x}{Vector.}
}
\value{
Numeric vector of same length as \code{.x}.
}
\description{
Safely convert to numeric
}
\examples{
as.numeric(factor("1.3"))
as_numeric(factor("1.3"))
}
