% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_vector.R
\name{as_vector}
\alias{as_vector}
\alias{as_vector.default}
\alias{as_vector.list}
\alias{as_vector.data.frame}
\alias{as_vector.data.table}
\title{As vector}
\usage{
as_vector(.x)

\method{as_vector}{default}(.x)

\method{as_vector}{list}(.x)

\method{as_vector}{data.frame}(.x)

\method{as_vector}{data.table}(.x)
}
\arguments{
\item{.x}{Object to convert.}
}
\description{
Convert a one-column or one-row data.table to an unnamed vector.
}
\examples{
data.frame(x = 1:5) \%>\% as_vector
data.frame(a = 1, b = 2) \%>\% as_vector
}
