% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce vectors.}
\usage{
coalesce(...)
}
\arguments{
\item{...}{Vectors of same length.

Copied from \href{https://stackoverflow.com/questions/19253820/how-to-implement-coalesce-efficiently-in-r}{Stackoverflow}.}
}
\value{
Vector of same length as input vectors.
}
\description{
Coalesce vectors.
}
\examples{
coalesce(c(NA, 1, NA), c(2, NA, NA))
}
