% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file.R
\name{current_file}
\alias{current_file}
\alias{current_dir}
\alias{setwd_current}
\title{Path to currently open document in RStudio}
\usage{
current_file()

current_dir()

setwd_current()
}
\value{
Character scalar with the path.
}
\description{
Path to currently open document in RStudio
}
\examples{
if (interactive()) {
  current_file()
}
}
