% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_date.R
\name{Date}
\alias{Date}
\alias{Datetime}
\title{Create a Date or Datetime vector}
\usage{
Date(length = 0L)

Datetime(length = 0L)
}
\description{
Equivalent to \code{character()}, \code{numeric()}, \code{integer()} etc.
}
\examples{
Date()
Date(5)
}
