% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\alias{infreq}
\title{Show most/least frequent values.}
\usage{
freq(.x, .n = 10, .prop = FALSE)

infreq(.x, .n = 10, .prop = FALSE)
}
\arguments{
\item{.x}{A vector passed on to \link{table}.}

\item{.n}{Integer scalar. Number of classes to show.}

\item{.prop}{Logical value. Show proportions?}
}
\value{
Table object.
}
\description{
These are convenient functions to show the most frequent or least frequent
values in a factor or character vector.
}
\examples{
iris[1:79, "Species"] \%>\% freq()
iris[1:79, "Species"] \%>\% infreq(2)
}
