% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq_table}
\alias{freq_table}
\title{Frequency table}
\usage{
freq_table(.x, .lump = TRUE, .cum = FALSE, .decreasing = TRUE,
  .digits = 3, ...)
}
\arguments{
\item{.x}{A vector of values, typical factor/character.}

\item{.lump}{Lump together small categories to "Other"?}

\item{.decreasing}{Logical value. Sort descending?}

\item{.digits}{Integer value. Number of digits to round proportions to.}

\item{...}{Arguments passed on to \code{\link[forcats:fct_lump]{forcats::fct_lump()}}.}
}
\value{
A data.table containing absolute and relative frequencies.
}
\description{
Show frequencies of categories in a data.table.
}
\examples{
iris$Species \%>\% freq_table()

if (require(nycflights13)) {
  nycflights13::flights$carrier \%>\% freq_table()
}
}
