% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{\%s\%}
\alias{\%s\%}
\title{Concatenate strings}
\usage{
.x \%s\% .y
}
\arguments{
\item{.x}{A character value.}

\item{.y}{A character value}
}
\value{
A character pasted together from .x and .y.
}
\description{
Infix operator for paste0.
}
\examples{
"Pasting " \%s\% "strings"
}
