% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unique.R
\name{is.unique}
\alias{is.unique}
\alias{is_unique}
\title{Is a vector unique?}
\usage{
is.unique(.x)

is_unique(.x)
}
\arguments{
\item{.x}{A vector.}
}
\value{
Logical scalar.
}
\description{
Is a vector unique?
}
