% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_input_empty}
\alias{is_input_empty}
\alias{is_empty}
\alias{is_empty_character}
\alias{is_empty_or_zero}
\alias{is_empty_or_false}
\alias{is_empty_or_na}
\title{Is input empty or falsy?}
\usage{
is_input_empty(.x)

is_empty(.x)

is_empty_character(.x)

is_empty_or_zero(.x)

is_empty_or_false(.x)

is_empty_or_na(.x)
}
\arguments{
\item{.x}{An object.}
}
\value{
Logical value.
}
\description{
Functions to check if inputs are NULL, zero length, "", 0 or FALSE.
Useful to check if an input inside a shiny app is empty.
}
\examples{
is_empty(NULL)
is_empty(character(0))

is_empty_or_false(FALSE)
is_empty_or_zero(0)
is_empty_or_na(NA)
is_empty_character("")
}
