% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%unlike\%}
\alias{\%xlike\%}
\alias{\%setdiff\%}
\title{Value Matching}
\usage{
.x \%notin\% .y

.x \%unlike\% .y

.x \%xlike\% .y

.x \%setdiff\% .y
}
\arguments{
\item{.x}{Vector. Values to be matched.}

\item{.y}{Vector. Values to be matched against.}
}
\value{
A logical vector of the same length as \code{.x}.
}
\description{
Opposite of \code{\%in\%}.
\code{\%notin\%} returns \code{TRUE} if there is no match between \code{.x} and \code{.table}.
}
\examples{
1 \%in\% 1:3
1 \%notin\% 1:3
"a" \%notin\% c("b", "c")
library(data.table)
names(iris) \%like\% "Length"
names(iris) \%xlike\% "Length"
names(iris) \%unlike\% "Length"
names(iris) \%setdiff\% c("Species", "x")
}
