% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_values.R
\name{map_values}
\alias{map_values}
\title{Map values}
\usage{
map_values(.x, .from, .to, .verbose = TRUE)
}
\arguments{
\item{.x}{Vector to modify.}

\item{.from}{Vector of values that should be replaced in \code{.x}.}

\item{.to}{Vector of new replacement values.}

\item{.verbose}{Logical scalar.}
}
\value{
A vector.
}
\description{
Replace values in a vector by matching values in a lookup vector.
This function is similar to \link[plyr:mapvalues]{plyr::mapvalues}.
}
\examples{
.x = c("Plebejus optilete", "Aglais io", "Plebejus optilete")
.from = c("Plebejus optilete", "Argynnis aglaja")
.to = c("Agriades optilete", "Speyeria aglaja")
map_values(.x, .from, .to)
}
