% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{zeros}
\alias{zeros}
\alias{ones}
\alias{random}
\title{Generate matrices}
\usage{
zeros(.nrow, .ncol)

ones(.nrow, .ncol)

random(.nrow, .ncol)
}
\arguments{
\item{.nrow}{Integer scalar. Number of rows.}

\item{.ncol}{Integer scalar. Number of columns. If missing will default to
the number of rows, i.e. a symmetrical matrix is created.}
}
\description{
Quick and easy generation of matrices.
Inspired by numpy.zeros etc.
}
\details{
\itemize{
\item \code{zeros} will create a matrix with all values 0L.
\item \code{ones} will create a matrix with all values 1L.
\item \code{random} will create a matrix with all values randomly uniform.
}
}
\examples{
zeros(4)
ones(3, 5)
random(5)
}
