% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible.R
\name{possible}
\alias{possible}
\alias{safe}
\alias{quiet}
\title{Capture side effects}
\usage{
possible(.expression, .otherwise, ...)

safe(.expression, ...)

quiet(.expression)
}
\arguments{
\item{.expression}{R Code which can be run.}

\item{.otherwise}{Default value to use when an error occurs.}

\item{...}{Arguments passed on to \code{\link[purrr:safely]{purrr::safely()}} or \link[purrr:safely]{purrr::possibly()}.}
}
\value{
\code{safe}: Returns a list with
components \code{result} and \code{error}. One value is always \code{NULL}.

\code{quiet}: Returns a list with components
\code{result}, \code{output}, \code{messages} and \code{warnings}.

\code{possible}: Uses a default value (\code{.otherwise})
whenever an error occurs.
}
\description{
These functions are equivalent to their purrr equivalents, but actually
run the code instead of returning a function.
}
\examples{
safe(log(10))
safe(log("a"))

possible(log(10), NA)
possible(log("e"), NA)

quiet(log(10))
quiet(log("a"))
}
\seealso{
\code{\link[purrr:safely]{purrr::safely()}}
}
