% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dt.R
\name{split_dt}
\alias{split_dt}
\title{Split a data.table into groups of similar length}
\usage{
split_dt(.Data, .n = NULL, .size = NULL)
}
\arguments{
\item{.Data}{A data.table.}

\item{.n}{Integer scalar. Number of groups.}

\item{.size}{Integer scalar. Number of observations per group.}
}
\description{
Split a data.table into groups of similar length
}
\examples{
library(data.table)
x <- split_dt(iris \%>\% as.data.table, 7)
lapply(x, NROW)

x <- split_dt(iris \%>\% as.data.table, .size = 50)
lapply(x, NROW)
}
