% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{str_collapse}
\alias{str_collapse}
\title{Collapse strings into one.}
\usage{
str_collapse(.x, .collapse = "")
}
\arguments{
\item{.x}{A character vector to collapse.}

\item{.collapse}{A separator string to be inserted between the elements of \code{.x}.}
}
\value{
A character vector of length 1.
}
\description{
Collapse strings into one.
}
\examples{
str_collapse(letters, ",")
}
