% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{str_discard}
\alias{str_discard}
\alias{str_keep}
\title{Keep / discard strings based on pattern}
\usage{
str_discard(.x, .pattern)

str_keep(.x, .pattern)
}
\arguments{
\item{.x}{A character vector to collapse.}

\item{.pattern}{A character value. Pattern to look for.}
}
\value{
A character vector.
}
\description{
Keep / discard strings based on pattern
}
\examples{
fruit <- c("apple", "banana", "pear", "pinapple")
str_keep(fruit, "apple")
str_discard(fruit, "apple")
}
