% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{str_split_v}
\alias{str_split_v}
\title{Split string into a vector.}
\usage{
str_split_v(.x, .pattern)
}
\arguments{
\item{.x}{A character scalar to split.}

\item{.pattern}{A pattern to apply the split.}
}
\value{
A character vector.
}
\description{
Similar to \code{\link[stringr:str_split]{stringr::str_split()}}, but returns a character vector,
instead of a list or matrix.
}
\examples{
str_split_v(c("a,b,c"), ",")
}
