% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize data.frame}
\usage{
summarize(.Data)
}
\arguments{
\item{.Data}{A data.frame or data.table.}
}
\description{
Summarize data.frame
}
\examples{
.Data <- data.table(a = letters[1:3],
                    b = 1:12,
                    c = runif(12),
                    d = c(NA, 1:11),
                    e = seq_date("2019-01-01", .length.out = 12))
summarize(.Data)

summarize(iris)
if (require(nycflights13)) {
  nycflights13::flights \%>\% summarize()
}
}
